move to new testing structure
