+++ tests/Integration/Managers/CustomizedManagerTest.php
@@ -1,0 +1,17 @@
+<?php
+namespace Michaels\Manager\Test\Integration\Managers;
+
+use Michaels\Manager\Manager;
+use Michaels\Manager\Test\Scenarios\InitManagerScenario;
+use Michaels\Manager\Test\Scenarios\ManagesItemsScenario;
+use Michaels\Manager\Test\Stubs\CustomizedManagerStub;
+
+/**
+ * Tests customized implementations of Manager.
+ * Does NOT test ManagesItemsTrait api.
+ */
+class CustomizedManagerTest extends \PHPUnit_Framework_TestCase {
+
+    use ManagesItemsScenario, InitManagerScenario; // Pull in all the tests for ManagesItemsTrait, for integration testing
+}
+
