+++ tests/Unit/Decoders/YamlDecoderTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace Michaels\Manager\Test;
+namespace Michaels\Manager\Test\Unit\Decoders;
 
 use Michaels\Manager\Decoders\YamlDecoder;
 use Symfony\Component\Yaml\Yaml;
