+++ tests/bootstrap.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once __DIR__ . "/../vendor/autoload.php";
+
+$GLOBALS['test_config'] = [
+    'test_dir' => __DIR__,
+];
